requirements.txt
beautifulsoup4==4.4.1
cffi==1.5.2
Flask==0.10.1
Flask-Autodoc==0.1.2
Flask-Misaka==0.4.1
Flask-SQLAlchemy==2.1
gunicorn==19.4.5
itsdangerous==0.24
Jinja==1.2
Jinja2==2.8
MarkupSafe==0.23
misaka==2.0.0
numpy==1.11.0
psycopg2==2.6.1
pycparser==2.14
python-dateutil==2.5.2
requests==2.9.1
six==1.10.0
SQLAlchemy==1.0.12
SQLAlchemy-Utils==0.32.3
w3lib==1.14.2
Werkzeug==0.11.5
